Applied review comments

+++ api/resume/routes.py
@@ -2,6 +2,7 @@ from resume.error.error import error
 from resume.dummy.dummyapi import dummyApi
 from resume.extraction.resumeextraction import ResumeExtract
 from flask import Flask, jsonify, request, Blueprint
+from api.resume_parser.parser import ResumeDataExtractor
 from flask_cors import CORS, cross_origin
 import os
 import json
@@ -34,6 +35,25 @@ def index():
                     file.save(file.filename)
                     ext = ResumeExtract(fileName)
                     dataList.append(ext.get_data())
+                    os.remove(fileName)
+            return json.dumps(dataList)
+        except Exception as e:
+            return error(str(e.args), 415)
+
+@main.route('/api/v1/postResumeExp', methods=['POST'])
+@cross_origin()
+def index():
+    dataList = list()
+    if request.method == 'POST':
+        try:
+            if not request.files or request.files['file'].filename == '':
+                raise Exception("Select a file")
+            for file in request.files.getlist('file'):
+                fileName = file.filename
+                if fileName.endswith('.pdf'):
+                    file.save(file.filename)
+                    ext = ResumeDataExtractor(fileName)
+                    dataList.append(ext.parse_resume())
                     os.remove(fileName)
             return json.dumps(dataList)
         except Exception as e:
